feat: limit selectable traits by list state
