+++ src/components/filters/action-filter.tsx
@@ -2,7 +2,6 @@ import { useCallback } from "react";
 
 import { useStore } from "@/store";
 import {
-  selectActionOptions,
   selectActiveListFilter,
   selectMultiselectChanges,
 } from "@/store/selectors/lists";
@@ -11,17 +10,20 @@ import { isActionFilterObject } from "@/store/slices/lists.type-guards";
 import { assert } from "@/utils/assert";
 import { capitalize } from "@/utils/formatting";
 
+import { selectActionOptions } from "@/store/selectors/lists";
+import type { FilterProps } from "./filters.types";
 import { MultiselectFilter } from "./primitives/multiselect-filter";
 
-export function ActionFilter({ id }: { id: number }) {
+export function ActionFilter({ id, resolvedDeck }: FilterProps) {
   const filter = useStore((state) => selectActiveListFilter(state, id));
+
   assert(
     isActionFilterObject(filter),
     `ActionFilter instantiated with '${filter?.type}'`,
   );
 
   const changes = selectMultiselectChanges(filter.value);
-  const options = useStore(selectActionOptions);
+  const options = useStore((state) => selectActionOptions(state, resolvedDeck));
 
   const nameRenderer = useCallback((item: Coded) => capitalize(item.code), []);
