+++ src/components/filters/filters.types.ts
@@ -1,0 +1,6 @@
+import type { ResolvedDeck } from "@/store/lib/types";
+
+export type FilterProps = {
+  id: number;
+  resolvedDeck?: ResolvedDeck;
+};
