+++ src/components/filters/subtype-filter.tsx
@@ -1,19 +1,20 @@
 import { useStore } from "@/store";
-import {
-  selectActiveListFilter,
-  selectSubtypeChanges,
-  selectSubtypeOptions,
-} from "@/store/selectors/lists";
+import { selectActiveListFilter } from "@/store/selectors/lists";
 import { isSubtypeFilterObject } from "@/store/slices/lists.type-guards";
 import { assert } from "@/utils/assert";
 
+import {
+  selectSubtypeChanges,
+  selectSubtypeOptions,
+} from "@/store/selectors/lists";
 import type { SubtypeFilter as SubtypeFilterType } from "@/store/slices/lists.types";
 import { useCallback } from "react";
 import { Checkbox } from "../ui/checkbox";
 import { CheckboxGroup } from "../ui/checkboxgroup";
+import type { FilterProps } from "./filters.types";
 import { FilterContainer } from "./primitives/filter-container";
 
-export function SubtypeFilter({ id }: { id: number }) {
+export function SubtypeFilter({ id, resolvedDeck }: FilterProps) {
   const filter = useStore((state) => selectActiveListFilter(state, id));
 
   assert(
@@ -25,7 +26,9 @@ export function SubtypeFilter({ id }: { id: number }) {
     selectSubtypeChanges(state, filter.value),
   );
 
-  const options = useStore(selectSubtypeOptions);
+  const options = useStore((state) =>
+    selectSubtypeOptions(state, resolvedDeck),
+  );
 
   const setFilterValue = useStore((state) => state.setFilterValue);
   const setFilterOpen = useStore((state) => state.setFilterOpen);
