+++ src/store/selectors/lists.ts
@@ -1,26 +1,227 @@
 import { createSelector } from "reselect";
 
-import { ASSET_SLOT_ORDER, FACTION_ORDER, SKILL_KEYS } from "@/utils/constants";
-import { capitalize, formatTabooSet } from "@/utils/formatting";
+import type { Filter } from "@/utils/fp";
+import { and, not, or } from "@/utils/fp";
+import { isEmpty } from "@/utils/is-empty";
 
+import { splitMultiValue } from "@/utils/card-utils";
+import { ASSET_SLOT_ORDER, FACTION_ORDER, SKILL_KEYS } from "@/utils/constants";
+import { createCustomEqualSelector } from "@/utils/custom-equal-selector";
+import { capitalize, formatTabooSet } from "@/utils/formatting";
+import { time, timeEnd } from "@/utils/time";
+import { applyCardChanges } from "../lib/card-edits";
+import { getAdditionalDeckOptions } from "../lib/deck-validation";
 import {
+  filterActions,
+  filterAssets,
+  filterCost,
+  filterEncounterCode,
+  filterFactions,
+  filterInvestigatorAccess,
+  filterInvestigatorWeaknessAccess,
+  filterLevel,
+  filterMythosCards,
+  filterOwnership,
+  filterPackCode,
+  filterProperties,
+  filterSkillIcons,
+  filterSubtypes,
+  filterTabooSet,
+  filterTraits,
+  filterType,
+} from "../lib/filtering";
+import { getGroupedCards } from "../lib/grouping";
+import { resolveCardWithRelations } from "../lib/resolve-card";
+import { applySearch } from "../lib/searching";
+import {
+  makeSortFunction,
   sortAlphabetical,
   sortByName,
   sortedEncounterSets,
 } from "../lib/sorting";
-import type { ResolvedDeck } from "../lib/types";
+import { type ResolvedDeck, isResolvedDeck } from "../lib/types";
 import type { Card, Cycle, Pack } from "../services/queries.types";
 import type { StoreState } from "../slices";
 import type {
   AssetFilter,
   CostFilter,
   LevelFilter,
+  List,
   MultiselectFilter,
+  OwnershipFilter,
   PropertiesFilter,
   SelectFilter,
   SkillIconsFilter,
   SubtypeFilter,
 } from "../slices/lists.types";
+import type { LookupTables } from "../slices/lookup-tables.types";
+import type { Metadata } from "../slices/metadata.types";
+import type { SettingsState } from "../slices/settings.types";
+
+export type CardGroup = {
+  type: string;
+  key: string;
+};
+
+export type ListState = {
+  key: string;
+  groups: CardGroup[];
+  cards: Card[];
+  groupCounts: number[];
+  totalCardCount: number;
+};
+
+function makeUserFilter(
+  metadata: Metadata,
+  lookupTables: LookupTables,
+  list: List,
+  settings: SettingsState,
+  targetDeck?: "slots" | "extraSlots" | "both",
+) {
+  const filters: Filter[] = [];
+
+  if (!list.filtersEnabled) return and(filters);
+
+  list.filters.forEach((_, id) => {
+    const filterValue = list.filterValues[id];
+    if (!filterValue) return;
+
+    switch (filterValue.type) {
+      case "action": {
+        const value = filterValue.value as MultiselectFilter;
+        if (value.length) {
+          filters.push(filterActions(value, lookupTables.actions));
+        }
+        break;
+      }
+
+      case "asset": {
+        const value = filterValue.value as AssetFilter;
+        const filter = filterAssets(value, lookupTables);
+        if (filter) filters.push(filter);
+        break;
+      }
+
+      case "cost": {
+        const value = filterValue.value as CostFilter;
+        if (value.range) filters.push(filterCost(value));
+        break;
+      }
+
+      case "encounterSet": {
+        const value = filterValue.value as MultiselectFilter;
+        if (value.length) filters.push(filterEncounterCode(value));
+        break;
+      }
+
+      case "faction": {
+        const value = filterValue.value as MultiselectFilter;
+        if (value.length) filters.push(filterFactions(value));
+        break;
+      }
+
+      case "investigator": {
+        const value = filterValue.value as string | undefined;
+
+        if (value) {
+          const filter = [];
+          const accessFilter = filterInvestigatorAccess(
+            metadata.cards[value],
+            lookupTables,
+            {
+              targetDeck,
+            },
+          );
+          const weaknessFilter = filterInvestigatorWeaknessAccess(
+            metadata.cards[value],
+            lookupTables,
+            { targetDeck },
+          );
+
+          if (accessFilter) filter.push(accessFilter);
+          if (weaknessFilter) filter.push(weaknessFilter);
+
+          filters.push(or(filter));
+        }
+
+        break;
+      }
+
+      case "level": {
+        const value = filterValue.value as LevelFilter;
+        if (value.range) filters.push(filterLevel(value));
+        break;
+      }
+
+      case "ownership": {
+        const value = filterValue.value as OwnershipFilter;
+        if (value !== "all") {
+          filters.push((card: Card) => {
+            const ownership = filterOwnership(
+              card,
+              metadata,
+              lookupTables,
+              settings.collection,
+              settings.showAllCards,
+            );
+
+            return value === "owned" ? ownership : !ownership;
+          });
+        }
+
+        break;
+      }
+
+      case "pack": {
+        const value = filterValue.value as MultiselectFilter;
+        if (value.length) {
+          const filter = filterPackCode(value, metadata, lookupTables);
+          if (filter) filters.push(filter);
+        }
+        break;
+      }
+
+      case "properties": {
+        const value = filterValue.value as PropertiesFilter;
+        filters.push(filterProperties(value, lookupTables));
+        break;
+      }
+
+      case "skillIcons": {
+        const value = filterValue.value as SkillIconsFilter;
+        filters.push(filterSkillIcons(value));
+        break;
+      }
+
+      case "subtype": {
+        const value = filterValue.value as SubtypeFilter;
+        filters.push(filterSubtypes(value));
+        break;
+      }
+
+      case "tabooSet": {
+        const value = filterValue.value as number | undefined;
+        if (value != null) filters.push(filterTabooSet(value, metadata));
+        break;
+      }
+
+      case "trait": {
+        const value = filterValue.value as MultiselectFilter;
+        if (value.length)
+          filters.push(filterTraits(value, lookupTables.traits));
+        break;
+      }
+
+      case "type": {
+        const value = filterValue.value as MultiselectFilter;
+        if (value.length) filters.push(filterType(value));
+        break;
+      }
+    }
+  });
+
+  return filters.length ? and(filters) : undefined;
+}
 
 export const selectActiveList = (state: StoreState) => {
   const active = state.activeList;
@@ -40,38 +241,384 @@ export const selectActiveListFilter = createSelector(
   },
 );
 
-/**
- * Filter options
- */
+export const selectCanonicalTabooSetId = (
+  state: StoreState,
+  resolvedDeck?: ResolvedDeck,
+) => {
+  if (resolvedDeck?.taboo_id) {
+    return resolvedDeck.taboo_id;
+  }
 
-export const selectActionOptions = createSelector(
-  (state: StoreState) => state.lookupTables.actions,
-  (actionTable) => {
-    const actions = Object.keys(actionTable).map((code) => ({ code }));
-    actions.sort((a, b) => sortAlphabetical(a.code, b.code));
-    return actions;
+  const filters = selectActiveListFilters(state);
+  const filterId = filters.findIndex((f) => f === "tabooSet");
+
+  const filterValue = filterId
+    ? selectActiveListFilter(state, filterId)
+    : undefined;
+
+  if (typeof filterValue?.value === "number") return filterValue.value;
+
+  return state.settings.tabooSetId;
+};
+
+// This selector uses a custom equality check that avoid re-creation on every deck change.
+// Deck access is only affected by a few subset of deck changes:
+// 1. The deck changes.
+// 2. The taboo that the deck uses changes.
+// 3. An investigator side changes.
+// 4. Cards that change deckbuilding rules (i.e. On Your Own, Versatile...) are added or removed.
+// 5. Customizations change, some options change card properties.
+// 6. Investigator option selections change.
+// 7. Deck card pool changes.
+// 8. Sealed deck chanees.
+const deckAccessEqualSelector = createCustomEqualSelector((a, b) => {
+  if (isResolvedDeck(a) && isResolvedDeck(b)) {
+    return (
+      a.id === b.id && // (1)
+      a.taboo_id === b.taboo_id && // 2
+      a.investigatorFront.card.code === b.investigatorFront.card.code && // 3
+      a.investigatorBack.card.code === b.investigatorBack.card.code && // 3
+      JSON.stringify(getAdditionalDeckOptions(a)) ===
+        JSON.stringify(getAdditionalDeckOptions(b)) && // 4
+      JSON.stringify(a.customizations) === JSON.stringify(b.customizations) && // 5
+      JSON.stringify(a.selections) === JSON.stringify(b.selections) && // 6
+      JSON.stringify(a.metaParsed.card_pool) ===
+        JSON.stringify(b.metaParsed.card_pool) && // 7
+      a.metaParsed.sealed_deck === b.metaParsed.sealed_deck // 8
+    );
+  }
+
+  // biome-ignore lint/suspicious/noDoubleEquals: we want a shallow equality check in this context.
+  return a == b;
+});
+
+// Mirrors the customization check above.
+const customizationsEqualSelector = createCustomEqualSelector((a, b) => {
+  return isResolvedDeck(a) && isResolvedDeck(b)
+    ? JSON.stringify(a.customizations) === JSON.stringify(b.customizations)
+    : // biome-ignore lint/suspicious/noDoubleEquals: we want a shallow equality check in this context.
+      a == b;
+});
+
+const selectDeckInvestigatorFilter = deckAccessEqualSelector(
+  (state: StoreState) => state.metadata,
+  (state: StoreState) => state.lookupTables,
+  (_: StoreState, resolvedDeck?: ResolvedDeck) => resolvedDeck,
+  (
+    _: StoreState,
+    __?: ResolvedDeck,
+    targetDeck?: "slots" | "extraSlots" | "both",
+  ) => targetDeck,
+  (state: StoreState) => state.ui.showUnusableCards,
+  (metadata, lookupTables, resolvedDeck, targetDeck, showUnusableCards) => {
+    if (!resolvedDeck) return undefined;
+
+    const investigator = resolvedDeck.investigatorBack.card;
+    if (!investigator) return undefined;
+
+    if (showUnusableCards) {
+      return and([
+        not(filterType(["investigator", "story", "location"])),
+        filterMythosCards,
+        (card: Card) =>
+          !lookupTables.relations.bonded[card.code] &&
+          (card?.xp != null ||
+            !card.restrictions ||
+            card.restrictions?.investigator?.[card.code]),
+      ]);
+    }
+
+    const ors = [];
+
+    const investigatorFilter = filterInvestigatorAccess(
+      investigator,
+      lookupTables,
+      {
+        additionalDeckOptions: getAdditionalDeckOptions(resolvedDeck),
+        selections: resolvedDeck.selections,
+        targetDeck,
+      },
+    );
+
+    const weaknessFilter = filterInvestigatorWeaknessAccess(
+      investigator,
+      lookupTables,
+      {
+        targetDeck,
+      },
+    );
+
+    if (investigatorFilter) ors.push(investigatorFilter);
+    if (weaknessFilter) ors.push(weaknessFilter);
+
+    const investigatorAccessFilter = or(ors);
+
+    const cardPool = resolvedDeck.cardPool;
+    const sealedDeck = resolvedDeck.sealedDeck?.cards;
+
+    if (!cardPool?.length && !sealedDeck) {
+      return investigatorAccessFilter;
+    }
+    const ands = [investigatorAccessFilter];
+
+    if (cardPool) {
+      const packFilter = filterPackCode(cardPool, metadata, lookupTables);
+      if (packFilter) {
+        ands.push(or([packFilter, (c) => c.xp == null]));
+      }
+    }
+
+    if (sealedDeck) {
+      const sealedMap = sealedDeck.reduce(
+        (acc, curr) => {
+          acc[curr] = true;
+          return acc;
+        },
+        {} as Record<string, boolean>,
+      );
+      const sealedFilter = (c: Card) => !!sealedMap[c.code];
+      ands.push(or([sealedFilter, (c) => c.xp == null]));
+    }
+
+    return and(ands);
   },
 );
 
+const selectResolvedDeckCustomizations = customizationsEqualSelector(
+  (_: StoreState, resolvedDeck?: ResolvedDeck) => resolvedDeck,
+  (resolvedDeck) => resolvedDeck?.customizations,
+);
+
+const selectBaseListCards = createSelector(
+  (state: StoreState) => state.metadata,
+  (state: StoreState) => selectActiveList(state)?.systemFilter,
+  selectDeckInvestigatorFilter,
+  selectCanonicalTabooSetId,
+  selectResolvedDeckCustomizations,
+  (
+    metadata,
+    systemFilter,
+    deckInvestigatorFilter,
+    tabooSetId,
+    customizations,
+  ) => {
+    if (isEmpty(metadata.cards)) {
+      console.warn("player cards selected before store is initialized.");
+      return undefined;
+    }
+
+    time("select_base_list_cards");
+
+    let filteredCards = Object.values(metadata.cards);
+
+    // filters can be impacted by card changes, apply them now.
+    if (tabooSetId || customizations) {
+      filteredCards = filteredCards.map((c) =>
+        applyCardChanges(c, metadata, tabooSetId, customizations),
+      );
+    }
+
+    // apply system filter early to cut down on # of cards that need to be processed.
+    if (systemFilter) filteredCards = filteredCards.filter(systemFilter);
+
+    if (deckInvestigatorFilter) {
+      filteredCards = filteredCards.filter(deckInvestigatorFilter);
+    }
+
+    timeEnd("select_base_list_cards");
+    return filteredCards;
+  },
+);
+
+export const selectListCards = createSelector(
+  (state: StoreState) => state.metadata,
+  (state: StoreState) => state.lookupTables,
+  (state: StoreState) => state.settings,
+  selectActiveList,
+  selectBaseListCards,
+  (metadata, lookupTables, settings, activeList, _filteredCards) => {
+    if (!_filteredCards || !activeList) return undefined;
+
+    time("select_list_cards");
+    let filteredCards = _filteredCards;
+
+    // apply search after initial filtering to cut down on search operations.
+    if (activeList.search.value) {
+      filteredCards = applySearch(activeList.search, filteredCards, metadata);
+    }
+
+    // this is the count of cards that a search would have matched before user filters are taken into account.
+    const totalCardCount = filteredCards.length;
+
+    // apply user filters.
+    const filter = makeUserFilter(metadata, lookupTables, activeList, settings);
+    if (filter) filteredCards = filteredCards.filter(filter);
+
+    const cards: Card[] = [];
+    const groups: CardGroup[] = [];
+    const groupCounts: number[] = [];
+
+    for (const group of getGroupedCards(
+      activeList.display.grouping,
+      filteredCards,
+      makeSortFunction(activeList.display.sorting, metadata),
+      metadata,
+    )) {
+      cards.push(...group.cards);
+
+      groups.push({
+        key: group.key,
+        type: group.type,
+      });
+
+      groupCounts.push(group.cards.length);
+    }
+
+    timeEnd("select_list_cards");
+
+    return {
+      key: activeList.key,
+      groups,
+      cards,
+      groupCounts,
+      totalCardCount,
+    } as ListState;
+  },
+);
+
+export const selectCardRelationsResolver = createSelector(
+  (state: StoreState) => state.metadata,
+  (state: StoreState) => state.lookupTables,
+  (metadata, lookupTables) => {
+    return (code: string) => {
+      // for the current use case (investigator signatures), customizations and taboo are irrelevant.
+      return resolveCardWithRelations(
+        metadata,
+        lookupTables,
+        code,
+        undefined,
+        undefined,
+        true,
+      );
+    };
+  },
+);
+
+/**
+ * Utilities
+ */
+
+const selectListFilterProperties = createSelector(
+  (state: StoreState) => state.lookupTables.actions,
+  selectBaseListCards,
+  (actionTable, cards) => {
+    time("select_card_list_properties");
+
+    const cost = { min: 0, max: 0 };
+    const health = { min: 0, max: 0 };
+    const sanity = { min: 0, max: 0 };
+
+    const actions: Set<string> = new Set();
+    const traits: Set<string> = new Set();
+    const subtypes: Set<string> = new Set();
+    const types: Set<string> = new Set();
+
+    if (cards) {
+      for (const card of cards) {
+        types.add(card.type_code);
+
+        if (card.cost != null && card.cost >= 0) {
+          cost.min = Math.min(cost.min, card.cost);
+          cost.max = Math.max(cost.max, card.cost);
+        }
+
+        if (card.health != null && card.health >= 0) {
+          health.min = Math.min(health.min, card.health);
+          health.max = Math.max(health.max, card.health);
+        }
+
+        if (card.sanity != null && card.sanity >= 0) {
+          sanity.min = Math.min(sanity.min, card.sanity);
+          sanity.max = Math.max(sanity.max, card.sanity);
+        }
+
+        if (card.subtype_code) {
+          subtypes.add(card.subtype_code);
+        }
+
+        for (const trait of splitMultiValue(card.real_traits)) {
+          traits.add(trait);
+        }
+
+        for (const trait of splitMultiValue(card.real_back_traits)) {
+          traits.add(trait);
+        }
+      }
+
+      for (const [key, table] of Object.entries(actionTable)) {
+        for (const card of cards) {
+          if (actions.has(key)) {
+            break;
+          }
+
+          if (table[card.code]) {
+            actions.add(key);
+          }
+        }
+      }
+    }
+
+    timeEnd("select_card_list_properties");
+
+    return {
+      actions,
+      cost,
+      health,
+      sanity,
+      traits,
+      subtypes,
+      types,
+    };
+  },
+);
+
+/**
+ * Actions
+ */
+
+export const selectActionOptions = createSelector(
+  selectListFilterProperties,
+  ({ actions }) => {
+    return Array.from(actions)
+      .sort()
+      .map((code) => ({ code }));
+  },
+);
+
+export const selectMultiselectChanges = (value: MultiselectFilter) => {
+  if (!value) return "";
+  return value.map(capitalize).join(" or ");
+};
+
+/**
+ * Asset
+ */
+
 export const selectAssetOptions = createSelector(
   (state: StoreState) => state.lookupTables,
-  (lookupTables) => {
-    const health = Object.keys(lookupTables.health).map((x) => +x);
-    const sanity = Object.keys(lookupTables.sanity).map((x) => +x);
-
+  selectListFilterProperties,
+  (lookupTables, filterProps) => {
     const uses = Object.keys(lookupTables.uses)
       .map((code) => ({ code }))
       .sort((a, b) => sortAlphabetical(a.code, b.code));
 
     const skillBoosts = SKILL_KEYS.filter((x) => x !== "wild");
 
-    health.sort();
-    sanity.sort();
     uses.sort();
 
     return {
-      health: { min: Math.max(health[0], 0), max: health[health.length - 1] },
-      sanity: { min: Math.max(sanity[0], 0), max: sanity[sanity.length - 1] },
+      health: filterProps.health,
+      sanity: filterProps.sanity,
       uses,
       slots: ASSET_SLOT_ORDER.map((code) => ({ code })),
       skillBoosts,
@@ -79,28 +626,89 @@ export const selectAssetOptions = createSelector(
   },
 );
 
-export const selectCostMinMax = (state: StoreState) => {
-  const costs = Object.keys(state.lookupTables.cost).map((x) =>
-    Number.parseInt(x, 10),
-  );
+export const selectAssetChanges = (value: AssetFilter) => {
+  const slot = value.slots.reduce((acc, key) => {
+    return !acc ? `Slot: ${capitalize(key)}` : `${acc} or ${key}`;
+  }, "");
 
-  if (costs.length < 2) {
-    throw new TypeError(
-      "selector {selectCostMinMax} expects store to contain metadata.",
-    );
+  const uses = value.uses.reduce((acc, key) => {
+    return !acc ? `Uses: ${capitalize(key)}` : `${acc} or ${key}`;
+  }, "");
+
+  const skillBoosts = value.skillBoosts.reduce((acc, key) => {
+    return !acc
+      ? `Skill boost: ${capitalize(key)}`
+      : `${acc} or ${capitalize(key)}`;
+  }, "");
+
+  let healthFilter = "";
+
+  if (value.health) {
+    let s = `${value.health[0]}`;
+    if (value.health[1] !== value.health[0]) {
+      s = `${s}-${value.health[1]}`;
+    }
+    healthFilter = `Health: ${s}`;
   }
 
-  costs.sort((a, b) => a - b);
+  let sanityFilter = "";
 
-  const min = 0; // arkhamdb data has some cards set to negative values.
-  const max = costs[costs.length - 1];
-  return [min, max];
+  if (value.sanity) {
+    let s = `${value.sanity[0]}`;
+
+    if (value.sanity[1] !== value.sanity[0]) {
+      s = `${s}-${value.sanity[1]}`;
+    }
+
+    sanityFilter = `Sanity: ${s}`;
+  }
+
+  return [slot, uses, skillBoosts, sanityFilter, healthFilter]
+    .filter((x) => x)
+    .join(", ");
 };
+
+/**
+ * Cost
+ */
+
+export const selectCostMinMax = createSelector(
+  selectListFilterProperties,
+  ({ cost }) => cost,
+);
+
+export const selectCostChanges = (value: CostFilter) => {
+  if (!value.range) return "";
+
+  let s = `${value.range[0]}`;
+  if (value.range[1] !== value.range[0]) s = `${s}-${value.range[1]}`;
+  if (value.even) s = `${s}, even`;
+  if (value.odd) s = `${s}, odd`;
+  if (value.x) s = `${s}, X`;
+
+  return s;
+};
+
+/**
+ * Encounter Set
+ */
 
 export const selectEncounterSetOptions = createSelector(
   (state: StoreState) => state.metadata,
   (metadata) => sortedEncounterSets(metadata),
 );
+
+export const selectEncounterSetChanges = createSelector(
+  (_: StoreState, value: MultiselectFilter) => value,
+  (state: StoreState) => state.metadata,
+  (value, metadata) => {
+    return value.map((id) => metadata.encounterSets[id].name).join(" or ");
+  },
+);
+
+/**
+ * Factions
+ */
 
 export const selectFactionOptions = createSelector(
   selectActiveList,
@@ -132,6 +740,10 @@ export const selectFactionOptions = createSelector(
   },
 );
 
+/**
+ * Investigator
+ */
+
 export const selectInvestigatorOptions = createSelector(
   (state: StoreState) => state.lookupTables,
   (state: StoreState) => state.metadata,
@@ -159,6 +771,35 @@ export const selectInvestigatorOptions = createSelector(
     return investigators;
   },
 );
+
+export const selectInvestigatorChanges = createSelector(
+  (_: StoreState, value: SelectFilter) => value,
+  (state: StoreState) => state.metadata,
+  (value, metadata) => {
+    if (!value) return "";
+    const card = metadata.cards[value];
+    return card
+      ? `${card.real_name}${card.parallel ? " (Parallel)" : ""}`
+      : value.toString();
+  },
+);
+
+/**
+ * Level
+ */
+
+export const selectLevelChanges = (value: LevelFilter) => {
+  if (!value.range) return undefined;
+  let s = `${value.range[0]}`;
+  if (value.range[1] !== value.range[0]) s = `${s}-${value.range[1]}`;
+  if (value.exceptional) s = `${s}, exceptional`;
+  if (value.nonexceptional) s = `${s}, nonexceptional`;
+  return s;
+};
+
+/**
+ * Packs
+ */
 
 type CycleWithPacks = (Cycle & {
   packs: Pack[];
@@ -220,13 +861,41 @@ export const selectPackOptions = createSelector(
   },
 );
 
-export function selectSubtypeOptions() {
-  return [
-    { code: "none", name: "None" },
-    { code: "weakness", name: "Weakness" },
-    { code: "basicweakness", name: "Basic weakness" },
-  ];
-}
+export const selectPackChanges = createSelector(
+  (_: StoreState, value: MultiselectFilter) => value,
+  (state: StoreState) => state.metadata,
+  (value, metadata) => {
+    if (!value) return "";
+    return value.map((id) => metadata.packs[id].real_name).join(" or ");
+  },
+);
+
+/**
+ * Properties
+ */
+
+export const selectPropertiesChanges = (value: PropertiesFilter) => {
+  return Object.entries(value).reduce((acc, [key, filterValue]) => {
+    if (!filterValue) return acc;
+    return !acc ? capitalize(key) : `${acc} and ${capitalize(key)}`;
+  }, "");
+};
+
+/**
+ * Skill Icons
+ */
+
+export const selectSkillIconsChanges = (value: SkillIconsFilter) => {
+  return Object.entries(value).reduce((acc, [key, val]) => {
+    if (!val) return acc;
+    const s = `${val}+ ${capitalize(key)}`;
+    return acc ? `${acc} and ${s}` : s;
+  }, "");
+};
+
+/**
+ * Taboo Set
+ */
 
 export const selectTabooSetOptions = (state: StoreState) => {
   const sets = Object.values(state.metadata.tabooSets);
@@ -243,169 +912,6 @@ export const selectTabooSetSelectOptions = (state: StoreState) => {
   }));
 };
 
-export const selectTraitOptions = createSelector(
-  selectActiveList,
-  (state: StoreState) => state.lookupTables.traitsByCardTypeSelection,
-  (activeList, traitTable) => {
-    if (!activeList) return [];
-    const types = Object.keys(traitTable[activeList.cardType]).map((code) => ({
-      code,
-    }));
-    types.sort((a, b) => sortAlphabetical(a.code, b.code));
-    return types;
-  },
-);
-
-export const selectTypeOptions = createSelector(
-  selectActiveList,
-  (state: StoreState) => state.metadata.types,
-  (state: StoreState) => state.lookupTables,
-  (activeList, typeTable, lookupTables) => {
-    if (!activeList) return [];
-
-    const types = Object.keys(
-      lookupTables.typesByCardTypeSelection[activeList.cardType],
-    ).map((type) => typeTable[type]);
-    types.sort((a, b) => sortAlphabetical(a.name, b.name));
-    return types;
-  },
-);
-
-/**
- * Filter changes
- */
-
-export const selectAssetChanges = (value: AssetFilter) => {
-  const slot = value.slots.reduce((acc, key) => {
-    return !acc ? `Slot: ${capitalize(key)}` : `${acc} or ${key}`;
-  }, "");
-
-  const uses = value.uses.reduce((acc, key) => {
-    return !acc ? `Uses: ${capitalize(key)}` : `${acc} or ${key}`;
-  }, "");
-
-  const skillBoosts = value.skillBoosts.reduce((acc, key) => {
-    return !acc
-      ? `Skill boost: ${capitalize(key)}`
-      : `${acc} or ${capitalize(key)}`;
-  }, "");
-
-  let healthFilter = "";
-
-  if (value.health) {
-    let s = `${value.health[0]}`;
-    if (value.health[1] !== value.health[0]) {
-      s = `${s}-${value.health[1]}`;
-    }
-    healthFilter = `Health: ${s}`;
-  }
-
-  let sanityFilter = "";
-
-  if (value.sanity) {
-    let s = `${value.sanity[0]}`;
-
-    if (value.sanity[1] !== value.sanity[0]) {
-      s = `${s}-${value.sanity[1]}`;
-    }
-
-    sanityFilter = `Sanity: ${s}`;
-  }
-
-  return [slot, uses, skillBoosts, sanityFilter, healthFilter]
-    .filter((x) => x)
-    .join(", ");
-};
-
-export const selectCostChanges = (value: CostFilter) => {
-  if (!value.range) return "";
-
-  let s = `${value.range[0]}`;
-  if (value.range[1] !== value.range[0]) s = `${s}-${value.range[1]}`;
-  if (value.even) s = `${s}, even`;
-  if (value.odd) s = `${s}, odd`;
-  if (value.x) s = `${s}, X`;
-
-  return s;
-};
-
-export const selectEncounterSetChanges = createSelector(
-  (_: StoreState, value: MultiselectFilter) => value,
-  (state: StoreState) => state.metadata,
-  (value, metadata) => {
-    return value.map((id) => metadata.encounterSets[id].name).join(" or ");
-  },
-);
-
-export const selectInvestigatorChanges = createSelector(
-  (_: StoreState, value: SelectFilter) => value,
-  (state: StoreState) => state.metadata,
-  (value, metadata) => {
-    if (!value) return "";
-    const card = metadata.cards[value];
-    return card
-      ? `${card.real_name}${card.parallel ? " (Parallel)" : ""}`
-      : value.toString();
-  },
-);
-
-export const selectLevelChanges = (value: LevelFilter) => {
-  if (!value.range) return undefined;
-  let s = `${value.range[0]}`;
-  if (value.range[1] !== value.range[0]) s = `${s}-${value.range[1]}`;
-  if (value.exceptional) s = `${s}, exceptional`;
-  if (value.nonexceptional) s = `${s}, nonexceptional`;
-  return s;
-};
-
-export const selectMultiselectChanges = (value: MultiselectFilter) => {
-  if (!value) return "";
-  return value.map(capitalize).join(" or ");
-};
-
-export const selectPropertiesChanges = (value: PropertiesFilter) => {
-  return Object.entries(value).reduce((acc, [key, filterValue]) => {
-    if (!filterValue) return acc;
-    return !acc ? capitalize(key) : `${acc} and ${capitalize(key)}`;
-  }, "");
-};
-
-export const selectPackChanges = createSelector(
-  (_: StoreState, value: MultiselectFilter) => value,
-  (state: StoreState) => state.metadata,
-  (value, metadata) => {
-    if (!value) return "";
-    return value.map((id) => metadata.packs[id].real_name).join(" or ");
-  },
-);
-
-export const selectSkillIconsChanges = (value: SkillIconsFilter) => {
-  return Object.entries(value).reduce((acc, [key, val]) => {
-    if (!val) return acc;
-    const s = `${val}+ ${capitalize(key)}`;
-    return acc ? `${acc} and ${s}` : s;
-  }, "");
-};
-
-export const selectSubtypeChanges = createSelector(
-  (_: StoreState, value: SubtypeFilter) => value,
-  (value) => {
-    const options = selectSubtypeOptions();
-
-    const enabled = options.filter(
-      ({ code }) => value[code as keyof SubtypeFilter],
-    );
-
-    if (enabled.length === 0) return "None";
-    if (enabled.length === options.length) return "";
-
-    return selectSubtypeOptions()
-      .filter(({ code }) => value[code as keyof SubtypeFilter])
-      .map(({ name }) => name)
-      .join(" or ");
-  },
-);
-
 export const selectTabooSetChanges = createSelector(
   (_: StoreState, value: SelectFilter) => value,
   (state: StoreState) => state.metadata,
@@ -416,25 +922,65 @@ export const selectTabooSetChanges = createSelector(
   },
 );
 
-export const selectCanonicalTabooSetId = (
-  state: StoreState,
-  resolvedDeck?: ResolvedDeck,
-) => {
-  if (resolvedDeck?.taboo_id) {
-    return resolvedDeck.taboo_id;
-  }
+/**
+ * Trait
+ */
 
-  const filters = selectActiveListFilters(state);
-  const filterId = filters.findIndex((f) => f === "tabooSet");
+export const selectTraitOptions = createSelector(
+  selectListFilterProperties,
+  ({ traits }) =>
+    Array.from(traits)
+      .sort()
+      .map((code) => ({ code })),
+);
 
-  const filterValue = filterId
-    ? selectActiveListFilter(state, filterId)
-    : undefined;
+/**
+ * Subtype
+ */
 
-  if (typeof filterValue?.value === "number") return filterValue.value;
-
-  return state.settings.tabooSetId;
+const subtypeLabels: Record<string, string> = {
+  none: "None",
+  weakness: "Weakness",
+  basicweakness: "Basic weakness",
 };
+
+export const selectSubtypeOptions = createSelector(
+  selectListFilterProperties,
+  ({ subtypes }) => {
+    return [
+      { code: "none", name: "None" },
+      ...Array.from(subtypes)
+        .sort()
+        .map((code) => ({ code, name: subtypeLabels[code] ?? "" })),
+    ];
+  },
+);
+
+export const selectSubtypeChanges = createSelector(
+  (_: StoreState, value: SubtypeFilter) => value,
+  (value) => {
+    const options = Object.entries(value);
+    const enabled = options.filter(([, value]) => !!value);
+
+    if (enabled.length === 0) return "None";
+    if (enabled.length === options.length) return "";
+
+    return enabled.map(([key]) => subtypeLabels[key]).join(" or ");
+  },
+);
+
+/**
+ * Type
+ */
+
+export const selectTypeOptions = createSelector(
+  selectListFilterProperties,
+  (state: StoreState) => state.metadata.types,
+  ({ types }, typeTable) =>
+    Array.from(types)
+      .sort()
+      .map((code) => typeTable[code]),
+);
 
 /**
  * Search
